Reworked code for the calendar cells to use a custom control which is based on the previously used button class.
This made a whole bunch of separate button layouts obsolete :-). The calendar cells are now marker with a "nicer" "today" marker (not just a plain circle any longer ;-)) and there are symbols to indicate the start of period and the calculated day of ovulation.

Partly painted cells for the transition between fertile and infertile days will now also be possible. In later versions there may also be a third symbol on the opposite edge to indicate stored entries for temperature etc. for this day.